Modification de l'arborescence du projet
